fixed dakota test script

+++ dakota-6.7/tests/run-tests.sh
@@ -1,9 +1,15 @@
 #!/bin/bash 
 
+# get input argument: image name
+if [ $# -eq 0 ]
+  then
+    echo "No arguments supplied. Must provide an image name to test against."
+    exit 1
+fi
+IMAGE=$1
 
 # run tests inside singleuser container
-#docker run --rm -u root -ti -v $(pwd)/entry.sh:/tmp/entry.sh \
 docker run --rm -u root -ti $IMAGE \
-"cd /tmp/dakota/build/test && ctest -j 4 -L FastTest -LE Diff && ctest -j 4 -L AcceptanceTest"
+-c  "cd /tmp/dakota/build/test && ctest -j 4 -L FastTest -LE Diff && ctest -j 4 -L AcceptanceTest"
